popravio proveru id-jeva prilikom dodavanja i editovanja komentara

+++ node/routes/comments.js
@@ -66,36 +66,51 @@ route.get('/comment/:id', (req, res) => {
 //sluzi samo da se promeni content
 route.put('/edit/:id', authMiddleware, (req, res) => {
     if(req.body === undefined){
-        res.status(400).send(new Error('Body empty error').sqlMessage);
+        res.status(400).send(new Error('Body empty error').message);
     } else {
-        if(req.user.user_id !== req.body.user_id){
-            return res.status(401).send(new Error('Unauthorized edit').sqlMessage);
-        }
         let {error} = Joi.validate(req.body, scheme);
         if (error) {
             res.status(400).send(error.details[0].message);
         }
         else {
-            let query = "update comment set content=? where id=?";
-            let formated = mysql.format(query, [req.body.content, req.params.id]);
-
-            pool.query(formated, (err, response) => {
+            let searchQuery = 'select * from comment where id=?';
+            let searchFormatted = mysql.format(searchQuery, [req.params.id]);
+            pool.query(searchFormatted, (err, rows) => {
                 if (err)
-                    res.status(500).send(err.sqlMessage);
+                    res.status(500).send(err.message);
                 else {
-                    query = 'select * from comment where id=?';
-                    formated = mysql.format(query, [req.params.id]);
-                    pool.query(formated, (err, rows) => {
-                        if (err)
-                            res.status(500).send(err.sqlMessage);
-                        else {
-                            if (rows[0] === undefined)
-                                res.status(400).send(new Error('Comment doesn\'t exist').sqlMessage);
-                            else {
-                                res.send(rows[0]);
-                            }
+                    if (rows[0] === undefined)
+                        res.status(400).send(new Error('Comment doesn\'t exist').nessage);
+                    else {
+                        if(req.user.user_id !== rows[0].user_id){
+                            console.log(rows[0].user_id)
+                            console.log(req.user.user_id)
+                            res.status(401).send(new Error('Unauthorized edit').message);
+                        }else{
+                            let query = "update comment set content=? where id=?";
+                            let formated = mysql.format(query, [req.body.content, req.params.id]);
+
+                            pool.query(formated, (err, response) => {
+                                if (err)
+                                    res.status(500).send(err.message);
+                                else {
+                                    query = 'select * from comment where id=?';
+                                    formated = mysql.format(query, [req.params.id]);
+                                    pool.query(formated, (err, rows) => {
+                                        if (err)
+                                            res.status(500).send(err.message);
+                                        else {
+                                            if (rows[0] === undefined)
+                                                res.status(400).send(new Error('Comment doesn\'t exist').message);
+                                            else {
+                                                res.send(rows[0]);
+                                            }
+                                        }
+                                    })
+                                }
+                            })
                         }
-                    })
+                    }
                 }
             })
         }
@@ -105,7 +120,7 @@ route.put('/edit/:id', authMiddleware, (req, res) => {
 route.post('/comment', authMiddleware, (req, res) => {
 
     if(req.user.user_id !== req.body.user_id){
-        return res.status(401).send(new Error('Unauthorized edit').sqlMessage);
+        return res.status(401).send(new Error('Unauthorized post').message);
     }
     let {error} = Joi.validate(req.body, scheme);
 
@@ -135,23 +150,27 @@ route.post('/comment', authMiddleware, (req, res) => {
 route.delete('/comment/:id', authMiddleware, (req, res) => {
     let query = 'select * from comment where id=?';
     let formatted = mysql.format(query, [req.params.id]);
-    let comment = null;
     pool.query(formatted, (err, rows) => {
         if(err){
-            return res.status(500).send(err.sqlMessage);
-        }else{
-            comment = rows[0];
-        }
-    });
-    query = 'delete from comment where id=?';
-    let formated = mysql.format(query, [req.params.id]);
-    pool.query(formated, (err, rows) => {
-        if (err)
             res.status(500).send(err.sqlMessage);
-        else {
-            res.status(200).send(comment);
+        }else{
+            let comment = rows[0];
+            if(comment.user_id !== req.user.user_id){
+                res.status(401).send(new Error('Unauthorized delete'));
+            }else {
+                query = 'delete from comment where id=?';
+                let formated = mysql.format(query, [req.params.id]);
+                pool.query(formated, (err, rows) => {
+                    if (err)
+                        res.status(500).send(err.sqlMessage);
+                    else {
+                        res.status(200).send(comment);
+                    }
+                });
+            }
         }
     });
+
 });
 
 module.exports = route;
